update spi receive api

+++ drivers/Inc/stm32f411xx_gpio_driver.h
@@ -1,7 +1,7 @@
 /*
- * stm32f411xx_gpio_driver.h
+ * stm32f411xx_gpio_driver_new.h
  *
- *  Created on: Oct 17, 2025
+ *  Created on: Nov 11, 2025
  *      Author: phanc
  */
 
@@ -20,7 +20,7 @@
 
 /* ========================= GPIO HANDLER - CONFIGURATION STRUCTURES ========================= */
 /* GPIO configuration structure */
-typedef struct 
+typedef struct
 {
     uint8_t GPIO_PinNumber;
     uint8_t GPIO_PinMode;           /* GPIO Pin Mode*/
@@ -32,7 +32,7 @@ typedef struct
 
 
 /* GPIO handler structure */
-typedef struct 
+typedef struct
 {
     /*Pointer to hold the base address of GPIO peripheral*/
     GPIO_RegDef_t *pGPIOx;       // This hold the  base address of the GPIO port to which the pin belongs
@@ -59,7 +59,7 @@ typedef struct
 #define GPIO_PIN_15          15
 
 /* GPIO Pin Mode*/
-#define GPIO_MODE_IN                    0   // Input 
+#define GPIO_MODE_IN                    0   // Input
 #define GPIO_MODE_OUT                   1   // General purpose output mode
 #define GPIO_MODE_ALTFN                 2   // Alternate Function mode
 #define GPIO_MODE_ANALOG                3   // Analog mode
@@ -104,8 +104,7 @@ void GPIO_ToggleOutputPin(GPIO_RegDef_t *pGPIOx, uint8_t PinNumber);
 
 /* IRQ configuration and ISR handling */
 void GPIO_IRQInterruptConfig(uint8_t IRQNumber, uint8_t EnorDi);
-void GPIO_IRQPriorityConfig(uint8_t IRQNumber, uint8_t IRQPriority);
+void GPIO_IRQPriorityConfig(uint8_t IRQNumber, uint32_t IRQPriority);
 void GPIO_IRQHandling(uint8_t PinNumber);
-
 
 #endif /* INC_STM32F411XX_GPIO_DRIVER_H_ */
